Implemented getMiddle function
This function takes a string as input and returns the middle character(s) of the string. If the string's length is odd, it returns the middle character. If the string's length is even, it returns the middle two characters.

This function enhances our string manipulation capabilities by providing a way to easily extract the middle character(s) from a string.